Diaryentry CRUD erstellt und Zugriff auf eigene User_id's eingeschränkt

+++ src/Controller/SeizureController.php
@@ -7,6 +7,7 @@ use App\Entity\User;
 use App\Form\SeizureType;
 use App\Repository\SeizureRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -57,6 +58,11 @@ class SeizureController extends AbstractController
      */
     public function show(Seizure $seizure): Response
     {
+        $user = $this->getUser();
+        if ($user->getId() != $seizure->getUser()->getId()) {
+            throw new AccessDeniedException('Zugriff verweigert');
+        }
+
         return $this->render('seizure/show.html.twig', [
             'seizure' => $seizure,
         ]);
@@ -67,6 +73,11 @@ class SeizureController extends AbstractController
      */
     public function edit(Request $request, Seizure $seizure): Response
     {
+        $user = $this->getUser();
+        if ($user->getId() != $seizure->getUser()->getId()) {
+            throw new AccessDeniedException('Zugriff verweigert');
+        }
+
         $form = $this->createForm(SeizureType::class, $seizure);
         $form->handleRequest($request);
 
@@ -89,6 +100,11 @@ class SeizureController extends AbstractController
      */
     public function delete(Request $request, Seizure $seizure): Response
     {
+        $user = $this->getUser();
+        if ($user->getId() != $seizure->getUser()->getId()) {
+            throw new AccessDeniedException('Zugriff verweigert');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$seizure->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($seizure);
